requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alluka==0.1.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorlog==6.7.0
frozenlist==1.3.1
heroku==0.1.4
hikari==2.0.0.dev111
hikari-lightbulb==2.2.4
hikari-tanjun==2.7.0a1
idna==3.4
multidict==6.0.2
python-dateutil==1.5
requests==2.28.1
urllib3==1.26.12
uvloop==0.17.0
yarl==1.8.1
